+++ docs/history.md
@@ -3,6 +3,7 @@
 ## Development version (unreleased)
 
 * FIX: Add history to cfradial1 output, and fix minor error in CfRadial1_Export.ipynb notebook({pull}`132`) by [@syedhamidali](https://github.com/syedhamidali)
+* FIX: align coordinates in backends ({pull}`139`) by [@kmuehlbauer](https://github.com/kmuehlbauer)
 
 ## 0.4.0 (2023-09-27)